backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
chromadb
groq
PyMuPDF
python-multipart
